update front n apiEndpoint
