+++ Frontend/my-app/.gitignore
@@ -1,17 +1,7 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
 /node_modules
-/.pnp
-.pnp.*
-.yarn/*
-!.yarn/patches
-!.yarn/plugins
-!.yarn/releases
-!.yarn/versions
 
-# testing
-/coverage
 
 # next.js
 /.next/
@@ -20,9 +10,7 @@
 # production
 /build
 
-# misc
-.DS_Store
-*.pem
+
 
 # debug
 npm-debug.log*
@@ -32,8 +20,6 @@ yarn-error.log*
 # env files (can opt-in for committing if needed)
 .env*
 
-# vercel
-.vercel
 
 # typescript
 *.tsbuildinfo
