+++ Frontend/my-app/src/app/addProduct/product-form/index.tsx
@@ -24,7 +24,7 @@ const AddProductForm: React.FC = () => {
     }
 
     try {
-      const response = await fetch('http://localhost:5000/api/products/addProduct', {
+      const response = await fetch(`${process.env.NEXT_PUBLIC_API_HOST}/api/products/addProduct`, {
         method: 'POST',
         body: formData,
       });
