+++ Frontend/my-app/src/types/index.ts
@@ -12,4 +12,13 @@ export interface CartItem {
   quantity: number;
   image: string
 }
-  
+ 
+export type Product = {
+  id: number;
+  title: string;
+  price: number;
+  category: string;
+  description: string;
+  image: string;
+  quantity: number;
+};
