+++ Frontend/my-app/src/types/product.ts
@@ -1,10 +1,10 @@
-// types.ts
-export type Product = {
-    id: number;
-    title: string;
-    price: number;
-    category: string;
-    description: string;
-    image: string;
-    quantity: number;
-};
+// // types.ts
+// export type Product = {
+//     id: number;
+//     title: string;
+//     price: number;
+//     category: string;
+//     description: string;
+//     image: string;
+//     quantity: number;
+// };
